requirements/server.txt
PyAudio
faster-whisper==0.6.0
--extra-index-url https://download.pytorch.org/whl/cu111
torch
torchaudio
websockets
pyannote.audio
onnxruntime==1.16.0
networkx
python-dotenv
intervaltree